+++ src/main/java/wooteco/subway/dto/respones/ErrorResponse.java
@@ -1,4 +1,4 @@
-package wooteco.subway.dto;
+package wooteco.subway.dto.respones;
 
 public class ErrorResponse {
