+++ src/main/java/wooteco/subway/ui/PathController.java
@@ -6,7 +6,7 @@ import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import wooteco.subway.dto.PathRequest;
-import wooteco.subway.dto.PathResponse;
+import wooteco.subway.dto.respones.PathResponse;
 import wooteco.subway.service.PathService;
 
 @RestController
