+++ src/main/java/wooteco/subway/ui/SubwayControllerAdvice.java
@@ -6,7 +6,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
-import wooteco.subway.dto.ErrorResponse;
+import wooteco.subway.dto.respones.ErrorResponse;
 import wooteco.subway.exception.BadRequestException;
 import wooteco.subway.exception.NotFoundException;
