+++ src/test/java/wooteco/subway/service/LineServiceTest.java
@@ -14,7 +14,7 @@ import wooteco.subway.domain.Line;
 import wooteco.subway.domain.Section;
 import wooteco.subway.domain.Station;
 import wooteco.subway.dto.LineRequest;
-import wooteco.subway.dto.LineResponse;
+import wooteco.subway.dto.respones.LineResponse;
 import wooteco.subway.dto.SectionRequest;
 import wooteco.subway.reopository.LineRepository;
 import wooteco.subway.reopository.SectionRepository;
